# src/track/dto/track.dto.ts
import { IsNumber, IsOptional, IsString } from '@nestjs/class-validator';

export class TrackDto {
  @IsString()
  readonly id: string;

  @IsString()
  name: string;

  @IsOptional()
  @IsString()
  artistId: string | null;

  @IsOptional()
  @IsString()
  albumId: string | null;

  @IsOptional()
  @IsNumber()
  duration?: number;
}